ttools: add weighted density shading mode
Working, but some changes to UI or presentation are probably required.

+++ ttools/src/main/uk/ac/starlink/ttools/plot2/layer/PixOutliner.java
@@ -69,7 +69,7 @@ public abstract class PixOutliner implements Outliner {
         final int xoff_;
         final int yoff_;
         final int[] counts_;
-        int pointCount_;
+        long pointCount_;
 
         /**
          * Constructor.
